% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_funs.R
\name{america_grid}
\alias{america_grid}
\title{Create a grid of points within the boundaries of the continental United States}
\usage{
america_grid(resolution)
}
\arguments{
\item{resolution}{The number of points along each axis of the grid.}
}
\value{
A matrix containing the latitude and longitude coordinates of the points within the
continental United States.
}
\description{
This function generates a grid of points within the boundaries of the continental United States.
The grid is defined by the specified \code{resolution}, which determines the number of points
along the latitude and longitude axes.
}
\examples{
# Generate a grid of points with a resolution of 100
grid <- america_grid(100)
}
