% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoning_estimate_units_per_parcel.R
\name{calculate_residential_units}
\alias{calculate_residential_units}
\title{Estimate the number and types of structures per parcel}
\usage{
calculate_residential_units(structures, parcels, zoning, acs = NULL)
}
\arguments{
\item{structures}{A dataset returned by \code{get_structure()}.}

\item{parcels}{A spatial (polygon) dataset.}

\item{zoning}{A spatial (polygon) zoning dataset.}

\item{acs}{Optionally, a non-spatial dataset, at the tract level, returned from \code{urbnindicators::compile_acs_data()}.}
}
\value{
The inputted parcels datasets with attributes describing estimated unit counts by unit type.
}
\description{
Estimate the number and types of structures per parcel
}
