% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climateapi-package.R
\docType{package}
\name{climateapi-package}
\alias{climateapi}
\alias{climateapi-package}
\title{climateapi: Climate (and adjacent) data and functions for the C&C practice area}
\description{
This packages combines a set of utilities for acquiring and processing climate and climate-adjacent datasets under a consistent API. It takes opinionated stances on how to manipulate raw data in an effort to produce standard workflows that enable project teams to devote more time to substantive analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ui-research.github.io/climateapi/}
}

}
\author{
\strong{Maintainer}: Will Curran-Groome \email{wcurrangroome@urban.org} (\href{https://orcid.org/0000-0001-6432-7561}{ORCID})

Authors:
\itemize{
  \item Kameron Lloyd
}

}
\keyword{internal}
