% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convert_delimited_to_parquet}
\alias{convert_delimited_to_parquet}
\title{Convert raw data to parquet to conserve memory / speed subsequent operations}
\usage{
convert_delimited_to_parquet(
  inpath,
  outpath = NULL,
  delimit_character = ",",
  subsetted_columns = NULL,
  dataset = NULL
)
}
\arguments{
\item{inpath}{The local path to read CSV data from.}

\item{outpath}{The local path to write parquet data to.}

\item{delimit_character}{The delimiting character of the raw data.}

\item{subsetted_columns}{The columns to include in the outputted parquet data.}

\item{dataset}{One of c("nfip_policies", "ihp_registrations"). If not null, this will be used to select the columns that are returned.}
}
\value{
Nothing. Parquet data are written to local path.
}
\description{
Convert raw data to parquet to conserve memory / speed subsequent operations
}
