% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_impacted_structures.R
\name{estimate_impacted_structures}
\alias{estimate_impacted_structures}
\title{Estimate counts of hazard-impacted structures by structure type}
\usage{
estimate_impacted_structures(boundaries, geography = "county")
}
\arguments{
\item{boundaries}{A POLYGON or MULTIPOLYGON object, or an sf::st_bbox()-style bbox.}

\item{geography}{The desired geography of the results. One of "tract" or "county".}
}
\value{
A dataframe comprising estimated counts of each structure type, at the specified \code{geography}, for all such geographic units intersecting the \code{boundaries} object.
}
\description{
Estimate counts of hazard-impacted structures by structure type
}
\examples{
\dontrun{
estimate_impacted_structures(
  geography = "tract",
  boundaries = tigris::states(cb = TRUE) \%>\% dplyr::filter(stringr::str_detect(NAME, "District")))
}
}
