% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_funs.R
\name{extract_data}
\alias{extract_data}
\title{Retrieve weather data for a given station id}
\usage{
extract_data(
  id,
  start = min(daily_data[daily_data$WBANNO == id, "LST_DATE"]),
  end = max(daily_data[daily_data$WBANNO == id, "LST_DATE"])
)
}
\arguments{
\item{id}{A unique numeric WBANNO id of the station to retrieve}

\item{start}{Character string in "&Y-%m-%d format representing the starting date
of the data to retrieve. Earliest possible date by default.}

\item{end}{Character string in "&Y-%m-%d format representing the
ending date of the data to retrieve. Latest possible date by default}
}
\value{
A dataframe with the following components:
\itemize{
\item{WBANNO} : {The station WBAN number.}
\item{LST_DATE} : {The Local Standard Time (LST) date of the observation.}
\item{CRX_VN} : {The version number of the station datalogger program that was
 in effect at the time of the observation.}
\item{LONGITUDE} : {Station longitude, using WGS-84.}
\item{LATITUDE} : {Station latitude, using WGS-84.}
\item{T_DAILY_MAX} : {Maximum air temperature, in degrees C.}
\item{T_DAILY_MIN} : {Minimum air temperature, in degrees C.}
\item{T_DAILY_MEAN} : {Mean air temperature, in degrees C, calculated using the
 typical historical approach: (T_DAILY_MAX + T_DAILY_MIN) / 2.}
\item{T_DAILY_AVG} : {Average air temperature, in degrees C.}
\item{P_DAILY_CALC} : {Total amount of precipitation, in mm.}
\item{SOLARAD_DAILY} : {Total solar energy, in MJ/meter^2, calculated from the
 hourly average global solar radiation rates and converted to energy by
 integrating over time.}
\item{SUR_TEMP_DAILY_TYPE} : {Type of infrared surface temperature measurement.}
\item{SUR_TEMP_DAILY_MAX} : {Maximum infrared surface temperature, in degrees C.}
\item{SUR_TEMP_DAILY_MIN} : {Minimum infrared surface temperature, in degrees C.}
\item{SUR_TEMP_DAILY_AVG} : {Average infrared surface temperature, in degrees C.}
\item{RH_DAILY_MAX} : {Maximum relative humidity, in \%.}
\item{RH_DAILY_MIN} : {Minimum relative humidity, in \%.}
\item{RH_DAILY_AVG} : {Average relative humidity, in \%.}
\item{SOIL_MOISTURE_5_DAILY} : {Average soil moisture, in fractional volumetric
 water content (m^3/m^3), at 5 cm below the surface.}
\item{SOIL_MOISTURE_10_DAILY} : {Average soil moisture, in fractional
 volumetric water content (m^3/m^3), at 10 cm below the surface.}
\item{SOIL_MOISTURE_20_DAILY} : {Average soil moisture, in fractional
 volumetric water content (m^3/m^3), at 20 cm below the surface.}
\item{SOIL_MOISTURE_50_DAILY} : {Average soil moisture, in fractional
 volumetric water content (m^3/m^3), at 50 cm below the surface.}
\item{SOIL_MOISTURE_100_DAILY} : {Average soil moisture, in fractional
 volumetric water content (m^3/m^3), at 100 cm below the surface.}
\item{SOIL_TEMP_5_DAILY} : {Average soil temperature, in degrees C, at 5 cm
 below the surface.}
\item{SOIL_TEMP_10_DAILY} : {Average soil temperature, in degrees C, at 10 cm
 below the surface.}
\item{SOIL_TEMP_20_DAILY} : {Average soil temperature, in degrees C, at 20 cm
 below the surface.}
\item{SOIL_TEMP_50_DAILY} : {Average soil temperature, in degrees C, at 50 cm
 below the surface.}
\item{SOIL_TEMP_100_DAILY} : {Average soil temperature, in degrees C, at 100 cm
 below the surface.}
 }
}
\description{
Retrieves the daily climate data from USCRN stations given a station id and
optionally a starting and ending date.
}
\examples{
 #Get weather data for station number 3047 (Monahans 6 TX)
 mon.dat<-extract_data(3047, start='2003-05-21', end='2003-05-27')
 mon.dat[,1:5]

}
