% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics_analysis.R
\name{format_qualtrics_metadata}
\alias{format_qualtrics_metadata}
\title{Prep Qualtrics Metadata}
\usage{
format_qualtrics_metadata(metadata, sections = c(), text_replace = "zzzzz")
}
\arguments{
\item{metadata}{A dataframe containing unprocessed metadata from the Qualtrics API}

\item{sections}{A named vector specifying the last question number in each survey section}

\item{text_replace}{A named character vector of regex patterns to replace in the metadata}
}
\value{
A dataframe of formatted metadata
}
\description{
Prep Qualtrics Metadata
}
