% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_box_path}
\alias{get_box_path}
\title{Get the path to the C&C Box folder}
\usage{
get_box_path()
}
\value{
A character string containing the full file path to the Climate and Communities (C&C) Box folder.
On Windows, returns "C:/Users/{username}/Box/METRO Climate and Communities Practice Area/github-repository".
On Mac, checks for Box at "/Users/{username}/Box" or "/Users/{username}/Library/CloudStorage/Box-Box",
using whichever exists. Throws an error if the Box folder cannot be found.
}
\description{
Get the path to the C&C Box folder
}
\examples{
\dontrun{
get_box_path()
}
}
