% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_fire_perimeters.R
\name{get_current_fire_perimeters}
\alias{get_current_fire_perimeters}
\title{Acquire current wildfire perimeters}
\usage{
get_current_fire_perimeters(
  geography = NULL,
  file_path = NULL,
  bbox = NULL,
  api = TRUE
)
}
\arguments{
\item{geography}{Included only for API consistency; this must be NULL.}

\item{file_path}{Included only for API consistency; this must be NULL.}

\item{bbox}{Optionally, an sf::st_bbox() object, or an object that can be converted to such.}

\item{api}{Included only for API consistency; this must be TRUE.}
}
\value{
An sf dataframe comprising perimeters of current wildfires. Columns include:
\describe{
\item{unique_id}{Unique identifier for each observation (generated).}
\item{incident_name}{Name of the fire incident (title case).}
\item{incident_size_acres}{Size of the fire in acres.}
\item{incident_short_description}{Brief description of the incident.}
\item{percent_contained}{Percent of fire contained (0-100).}
\item{identified_date}{Date/time the fire was discovered.}
\item{updated_date}{Date/time the record was last updated.}
\item{geometry}{Polygon geometry of the fire perimeter.}
}
}
\description{
Retrieves current wildfire perimeter data from the NIFC (National
Interagency Fire Center) via the Wildland Fire Interagency Geospatial Services
(WFIGS) API.
}
\details{
Data are from the NIFC WFIGS service. See
\url{https://data-nifc.opendata.arcgis.com/datasets/nifc::wfigs-interagency-fire-perimeters/about}.
}
\examples{
\dontrun{
get_current_fire_perimeters()
}
}
