% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fema_disaster_declarations_county.R
\name{get_fema_disaster_declarations_county}
\alias{get_fema_disaster_declarations_county}
\title{Get Major Disaster Declarations by County}
\usage{
get_fema_disaster_declarations_county(
  geography = "county",
  file_path = file.path(get_box_data_path(), "hazards", "fema", "disaster-declarations",
    "raw", "fema_disaster_declarations_county_2024_10_25.csv"),
  api = TRUE
)
}
\arguments{
\item{geography}{Included for API consistency only; must be \code{county}.}

\item{file_path}{The path (on Box) to the file containing the raw data.}

\item{api}{If TRUE (default), access data from the API. Else, read locally from \code{file_path}.}
}
\value{
A dataframe comprising Major Disaster Declarations by month by year by county.
}
\description{
Get Major Disaster Declarations by County
}
\examples{
\dontrun{
get_fema_disaster_declarations_county(
  api = TRUE)
}
}
