% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_geography_metadata}
\alias{get_geography_metadata}
\title{Get geography metadata about states or counties}
\usage{
get_geography_metadata(geography_type = c("state", "county"))
}
\arguments{
\item{geography_type}{One of c("state", "county").}
}
\value{
A data frame containing metadata about the specified geography type and area.
}
\description{
Get geography metadata about states or counties
}
