% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hazard_mitigation_assistance.R
\name{get_hazard_mitigation_assistance}
\alias{get_hazard_mitigation_assistance}
\title{Get Hazard Mitigation Assistance (HMA) Project Details}
\usage{
get_hazard_mitigation_assistance(
  file_path_old_grant_system = file.path(get_box_path(), "hazards", "fema",
    "hazard-mitigation-assistance", "raw",
    "HazardMitigationAssistanceProjects_2025_09_27.parquet"),
  file_path_new_grant_system = file.path(get_box_path(), "hazards", "fema",
    "hazard-mitigation-assistance", "raw", "HmaSubapplications_2025_09_27.parquet"),
  state_abbreviations = NULL
)
}
\arguments{
\item{file_path_old_grant_system}{The file path to raw data for HMA applications from the older grant-reporting system. These data are typically available from: https://www.fema.gov/openfema-data-page/hazard-mitigation-assistance-projects-v4}

\item{file_path_new_grant_system}{The file path to raw data for HMA applications from the newer (FEMA GO) grant-reporting system. These data are typically available from: https://www.fema.gov/openfema-data-page/hma-subapplications-v2}

\item{state_abbreviations}{NULL by default, in which case data are returned for all 51 states. Provide a vector of two-character USPS state abbreviations to obtain data for a sub-selection of states.}
}
\value{
A dataframe of project-county HMA application data, aggregated across both old and new grant reporting systems.
}
\description{
Get Hazard Mitigation Assistance (HMA) Project Details
}
\examples{
\dontrun{
get_hazard_mitigation_assistance()
}
}
