% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ihp_registrations.R
\name{get_ihp_registrations}
\alias{get_ihp_registrations}
\title{Get Individuals and Households Program (IHP) registrations}
\usage{
get_ihp_registrations(
  state_fips = NULL,
  file_name = "IndividualsAndHouseholdsProgramValidRegistrationsV2_2025_09_26.parquet",
  api = FALSE,
  outpath = NULL
)
}
\arguments{
\item{state_fips}{A character vector of two-letter state abbreviations. If NULL (default), return data for all 51 states. Otherwise return data for the specified states.}

\item{file_name}{The name (not the full path) of the Box file containing the raw data.}

\item{api}{If TRUE, query the API. If FALSE (default), read from disk.}

\item{outpath}{The path to save the parquet-formatted datafile. Applicable only when \code{api = FALSE}.}
}
\value{
A dataframe comprising IHP registrations
}
\description{
Get Individuals and Households Program (IHP) registrations
}
\examples{
\dontrun{
get_ihp_registrations(
   state_fips = "NJ",
   api = TRUE)
}
}
