% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lehd.R
\name{get_lehd}
\alias{get_lehd}
\title{Get LEHD Origin-Destination Employment Statistics (LODES) data
Returned data are from LODES Version 8, which is enumerated in 2020-vintage geometries.}
\usage{
get_lehd(
  lodes_type,
  jobs_type = "all",
  states,
  years,
  geography = "tract",
  state_part = "aux"
)
}
\arguments{
\item{lodes_type}{One of c("rac", "wac", "od"). "rac" = Residence Area Characteristics, where jobs are associated with employees' residences. "wac" = Workplace Area Characteristics, where jobs are associated with employees' workplaces. "od" = Origin-Destination data, where jobs are associated with both workers' residences and their workplaces.}

\item{jobs_type}{One of c("all", "primary"). Default is "all", which includes multiple jobs for workers with multiple jobs. "primary" includes only the highest-paying job per worker.}

\item{states}{A vector of state abbreviations.}

\item{years}{A vector of years.}

\item{geography}{One of c("block", "block group", "tract", "county", "state"). Default is "tract".}

\item{state_part}{One of c("main", "aux"). Default is "aux", which includes workers who reside outside of the state where they work.}
}
\value{
A tibble with one record per geography per year per job type. Attributes include total jobs and jobs by worker earnings, industry, and demographics; the origin-destination results have more limited demographics compared to the "wac" and "rac" results.
}
\description{
Get LEHD Origin-Destination Employment Statistics (LODES) data
Returned data are from LODES Version 8, which is enumerated in 2020-vintage geometries.
}
