% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nfip_claims.R
\name{get_nfip_claims}
\alias{get_nfip_claims}
\title{Access county-level data on NFIP policies}
\usage{
get_nfip_claims(
  county_geoids = NULL,
  file_name = "fima_nfip_claims_2025_06_08.parquet",
  api = FALSE
)
}
\arguments{
\item{county_geoids}{A character vector of five-digit county codes. NULL by default; must be non-NULL if \code{api = TRUE}.}

\item{file_name}{The name (not the full path) of the Box file containing the raw data.}

\item{api}{If TRUE, query the API. FALSE by default.}
}
\value{
A dataframe comprising county-level data on current NFIP policies
}
\description{
Access county-level data on NFIP policies
}
\examples{
\dontrun{
get_nfip_claims(
   county_geoids = c("01001", "01003"),
   api = TRUE)
}
}
