% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nfip_residential_penetration.R
\name{get_nfip_residential_penetration}
\alias{get_nfip_residential_penetration}
\title{Get the share of residential structures covered by NFIP}
\usage{
get_nfip_residential_penetration(
  states = NULL,
  file_name = "nfip_residential_penetration_rates_12_12_2025.csv"
)
}
\arguments{
\item{states}{NULL by default.}

\item{file_name}{The name (not full path) of the raw dataset.}
}
\value{
A tibble with the following columns:
\describe{
\item{state_name}{State name}
\item{couty_name}{County name}
\item{GEOID}{Five digits county FIPS code}
\item{year_data}{The year of the data}
\item{penetration_rate}{Share of all residential structures insured by NFIP}
\item{penetration_rate_sfha}{Share of all residential structures in the Special Flood Hazard Area insured by NFIP}
\item{contracts_in_force}{Residential NFIP contracts currently in effect ("in force")}
\item{contracts_in_force_sfha}{Residential NFIP contracts in the Special Flood Hazard Area currently in effect ("in force")}
\item{residential_structures}{The number of residential structures, derived from the 2022 National Structure Inventory}
\item{residential_structures_sfha}{The number of residential structures in the Special Flood Hazard Area, derived from the 2022 National Structure Inventory}
}
}
\description{
Get the share of residential structures covered by NFIP
}
\examples{
\dontrun{
get_nfip_residential_penetration()
}
}
