% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_openfema_cache_path}
\alias{get_openfema_cache_path}
\title{Get the path to the local OpenFEMA dataset cache}
\usage{
get_openfema_cache_path()
}
\value{
A character string containing the full path to the local cache of
OpenFEMA datasets, populated via \code{download_openfema_datasets()}. This
cache lives at the root of the user's Box folder (\code{data-cache/openfema}),
not under the C&C practice area subfolder returned by \code{get_box_path()}.
}
\description{
Get the path to the local OpenFEMA dataset cache
}
\examples{
\dontrun{
get_openfema_cache_path()
}
}
