% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_public_assistance.R
\name{get_public_assistance}
\alias{get_public_assistance}
\title{Get FEMA Public Assistance (PA) funding}
\usage{
get_public_assistance(
  file_path = file.path(get_box_path(), "hazards", "fema", "public-assistance", "raw",
    "public_assistance_funded_projects_county_2024_10_25.parquet"),
  state_abbreviations = NULL
)
}
\arguments{
\item{file_path}{The file path to the raw data file, preferably a .parquet file.}

\item{state_abbreviations}{A character vector of state abbreviations. NULL by default, which returns records for all 51 states.}
}
\value{
A dataframe of project-level funding requests and awards, along with variables that can be aggregated to the county level.
}
\description{
Project- and county-level data on PA funding over time
}
\examples{
\dontrun{
  get_public_assistance(state_abbreviations = "NJ")
}
}
