% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics_analysis.R
\name{get_qualtrics_metadata}
\alias{get_qualtrics_metadata}
\title{Access Qualtrics Metadata}
\usage{
get_qualtrics_metadata(
  metadata,
  question_name = NULL,
  survey_section = NULL,
  return_values = "text_sub"
)
}
\arguments{
\item{metadata}{The dataframe containing the Qualtrics metadata}

\item{question_name}{A regex pattern to match the question name(s)}

\item{survey_section}{A regex pattern to match the survey section(s)}

\item{return_values}{The name of the column (character) to be returned}
}
\value{
A character vector of the requested metadata
}
\description{
Access Qualtrics Metadata
}
