% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sheldus_county.R
\name{get_sheldus_county}
\alias{get_sheldus_county}
\title{Access temporal county-level SHELDUS hazard counts.}
\usage{
get_sheldus_county(
  geography = "county",
  file_path = file.path(get_box_data_path(), "hazards", "sheldus",
    "SHELDUS_22.0_01012916_01012023_AllCounties_CountyAggregate_YearMonth_2022USD",
    "UID14148f_AGG_A.csv"),
  api = FALSE
)
}
\arguments{
\item{geography}{Required -- at what geography should the data be aggregated? Must be \code{county}.}

\item{file_path}{The path from which to read the dataset if \code{api = FALSE}.}

\item{api}{Included for consistency; must be FALSE.}
}
\value{
A dataframe comprising month x year x county counts of hazard events.
}
\description{
Access temporal county-level SHELDUS hazard counts.
}
\examples{
\dontrun{
get_sheldus_county()
}
}
