% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_funs.R
\name{interpolate_data}
\alias{interpolate_data}
\title{Interpolate Data Using GpGp's Gaussian Process}
\usage{
interpolate_data(response, dat = daily_data, resolution = 100)
}
\arguments{
\item{response}{The name of the response variable to be interpolated.}

\item{dat}{The dataframe containing the data to be used for interpolation.
By default, it uses the daily_data dataframe.}

\item{resolution}{The resolution of the grid used for interpolation.
Default is 100.}
}
\value{
A dataframe containing the longitude, latitude, and interpolated
values of the response variable.
}
\description{
This function performs data interpolation using GpGp.It fits a GPR model to
the provided data (\code{dat}), predicts the values at specified locations
using the \code{america_grid} function, and returns a dataframe containing
the interpolated values along with their corresponding longitude and latitude
coordinates.
}
\examples{
# Interpolate temperature data
fake.dat <- data.frame(LONGITUDE = runif(35, 30, 40),
LATITUDE=runif(35, 30, 40),
means = runif(35, 30, 40))
interp <- interpolate_data('means', dat= fake.dat,resolution = 150)
}
