% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_sociodemographics_zones.R
\name{interpolate_sociodemographics_to_zones}
\alias{interpolate_sociodemographics_to_zones}
\title{Interpolate tract-level sociodemographic data to zoning polygons}
\usage{
interpolate_sociodemographics_to_zones(
  zones_sf,
  sociodemographic_tracts_sf = NULL,
  id_column,
  weights = "population"
)
}
\arguments{
\item{zones_sf}{A spatial (sf) dataset defining zoning districts.}

\item{sociodemographic_tracts_sf}{(optional) A tract level spatial (sf) dataset resulting from urbnindicators. If NULL, this function is run behind the scenes for appropriate tracts.}

\item{id_column}{The name of the column in zones_sf that identifies each unique combination of zoning regulations.}

\item{weights}{One of c("population", "housing"). The variable to be used as the weight in the interpolation.}
}
\value{
A spatial (sf) dataset comprising one observation for each level of id_column with interpolated values taken from sociodemographic_tracts_sf.
}
\description{
Interpolate tract-level sociodemographic data to zoning polygons
}
