% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_funs.R
\name{plot_predictions}
\alias{plot_predictions}
\title{Plot Predictions as Heatmap}
\usage{
plot_predictions(df)
}
\arguments{
\item{df}{A dataframe containing longitude, latitude, and predicted values.
Should be a return value of \code{interpolate_values}}
}
\value{
A ggplot2 object representing the heatmap plot.
}
\description{
This function creates a heatmap plot to visualize predicted values over a geographical area.
}
\examples{
# Create a dataframe
fake.dat <- data.frame(LONGITUDE = runif(35, 30, 40),
LATITUDE=runif(35, 30, 40),
means = runif(35, 30, 40))
interp <- interpolate_data('means', dat= fake.dat,resolution = 150)#' # Plot predictions
plot <- plot_predictions(interp)
}
