% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_analysis.R
\name{polygons_to_linestring}
\alias{polygons_to_linestring}
\title{Convert polygons into their component linestrings}
\usage{
polygons_to_linestring(.sf)
}
\arguments{
\item{.sf}{The spatial dataframe containing one or more polygons}
}
\value{
A simple feature collection of linestrings derived from the inputted polygons; all attributes are retained, and two new attributes--\code{polygon_id} and \code{line_id}--are prepended to the output
}
\description{
Convert polygons into their component linestrings
}
