% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nfip_policies_county.R
\name{raw_nfip_policies_to_parquet}
\alias{raw_nfip_policies_to_parquet}
\title{Convert raw data to parquet to conserve memory / speed subsequent operations.}
\usage{
raw_nfip_policies_to_parquet(
  inpath,
  outpath = NULL,
  delimit_character = ",",
  subsetted_columns = NULL
)
}
\arguments{
\item{inpath}{The local path to read CSV data from.}

\item{outpath}{The local path to write parquet data to.}

\item{delimit_character}{The delimiting character of the raw data.}

\item{subsetted_columns}{The columns to include in the outputted parquet data.}
}
\value{
Nothing. Parquet data are written to local path.
}
\description{
Convert raw data to parquet to conserve memory / speed subsequent operations.
}
