% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ipums_micro_cached.R
\name{read_ipums_micro_cached}
\alias{read_ipums_micro_cached}
\title{Read Cached IPUMS Microdata}
\usage{
read_ipums_micro_cached(
  filename,
  download_directory,
  extract_definition,
  refresh = FALSE
)
}
\arguments{
\item{filename}{name of the file (not the full path)--do not include a file extension}

\item{download_directory}{wherever the data will be downloaded to--must be a relative path}

\item{extract_definition}{the object resulting from define_extract_micro()}

\item{refresh}{if true, execute the API query, even if data are already stored locally}
}
\value{
A dataframe corresponding to the supplied extract_definition
}
\description{
Read Cached IPUMS Microdata
}
\examples{
\dontrun{
read_ipums_micro_cached(
  filename = "acs_insurance_race_2022_1yr_repweights",
  download_directory = "data",
  extract_definition = define_micro_extract(
    collection = "usa",
    description = "2022 ACS 1-year sample with replicate weights - insurance and race",
    samples = c("us2022a"),
    variables = list(
      "HCOVANY",
      var_spec("RACE", case_selections = c("1", "2")),
      "REPWT"),
  refresh = FALSE))
}

}
