% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stations}
\alias{stations}
\title{Information on U.S. Climate Reference Network Stations}
\format{
a dataframe with 237 rows and 5 columns
\describe{
\item{Station_Name}{The name of the USCRN Station, usually the City in
which it is located}
\item{State}{The two letter abbreviation of the State that the station is in}
\item{Station_ID}{The station WBAN number}
\item{Longitude}{Station longitude, using WGS-84.}
\item{Latitude}{Station latitude, using WGS-84}
}
}
\usage{
data(stations)
}
\description{
This dataset contains location and identification information on all 237
stations used to collect daily climate data across America by the USCRN
}
\keyword{datasets}
