% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_analysis.R
\name{subdivide_linestring}
\alias{subdivide_linestring}
\title{Subdivide a linestring into segments of a specified length}
\usage{
subdivide_linestring(line, max_length, crs = 5070)
}
\arguments{
\item{line}{A linestring or simple feature collection thereof}

\item{max_length}{The maximum length of each segment. Segments longer than this value will be subdivided; those that are below this threshold will be returned as-is.}

\item{crs}{The coordinate reference system to which the linestring should be transformed. Default is 5070.}
}
\value{
A spatial dataframe comprising linestrings below the \code{max_length} threshold, linked back to their input linestrings via a \code{line_id} attribute
}
\description{
Subdivide a linestring into segments of a specified length
}
