% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_trends.R
\name{temp_trends}
\alias{temp_trends}
\title{Finds the expected yearly change in average temperature in degrees celsius
for a given station.}
\usage{
temp_trends(id)
}
\arguments{
\item{id}{A unique numeric WBANNO id of the station to fit the model for}
}
\description{
Fits a linear model of average temperatures of USCRN stations to years
and returns its beta coefficient.
}
\examples{
 #Get expected temperature change for station number 3047 (Monahans 6 TX)
 temp_trends(3047)
}
