% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_funs.R
\name{yearly_cycle}
\alias{yearly_cycle}
\title{A function for estimating the yearly cycle for one station.}
\usage{
yearly_cycle(id)
}
\arguments{
\item{id}{A unique numeric WBANNO station ID of the station to retrieve}
}
\value{
A 365x2 dataframe with the following format:
\itemize{
\item{Day} : {Numbered day of the year (1-365)}
\item{Exp_Temp} : {Expected Temperature for the corresponding day}
}
}
\description{
Retrieves the yearly cycle for a given station id, where a yearly cycle is
the expected temperature on each day of the year.
}
\examples{
 #Get yearly cycle data for station 3047, Monohans 6 in TX
 mon6.cycle <-yearly_cycle(3047)
 head(mon6.cycle)

}
