% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_funs.R
\name{yearly_cycle_one_year}
\alias{yearly_cycle_one_year}
\title{A hidden function for estimating the yearly cycle for one station given a
starting and ending date to calculate means from.}
\usage{
yearly_cycle_one_year(
  id,
  start = min(daily_data[daily_data$WBANNO == id, "LST_DATE"]),
  end = max(daily_data[daily_data$WBANNO == id, "LST_DATE"])
)
}
\arguments{
\item{id}{A unique numeric WBANNO station ID of the station to retrieve}

\item{start}{First date you would like to calculate mean using}

\item{end}{Latest date you would like to use to calculate the mean}
}
\value{
A 365x2 dataframe with the following format:
\itemize{
\item{Day} : {Numbered day of the year (1-365)}
\item{Exp_Temp} : {Expected Temperature for the corresponding day}
}
}
\description{
Retrieves the yearly cycle for a given station and returns a dataframe of
the expected temperatures found using a cyclical linear model fitted to the
historical mean temperatures. Used in other functions so is not available
for public users
}
