% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_box_path}
\alias{get_box_path}
\title{Get the path to the C&C Box folder}
\usage{
get_box_path()
}
\value{
The filepath to the C&C Box folder
}
\description{
Get the path to the C&C Box folder
}
\examples{
\dontrun{
get_box_path()
}
}
