% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_business_patterns.R
\name{get_business_patterns}
\alias{get_business_patterns}
\title{Obtain County Business Patterns (CBP) Estimates per County}
\usage{
get_business_patterns()
}
\value{
A tibble with data on county-level employees, employers, and aggregate annual payrolls by industry and employer size
}
\description{
Obtain County Business Patterns (CBP) Estimates per County
}
\examples{
\dontrun{
get_business_patterns()
}
}
