% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_fire_perimeters.R
\name{get_current_fire_perimeters}
\alias{get_current_fire_perimeters}
\title{Acquire wildfire perimeters}
\usage{
get_current_fire_perimeters(
  geography = NULL,
  file_path = NULL,
  bbox = NULL,
  api = TRUE
)
}
\arguments{
\item{geography}{Included only for API consistency; this must be NULL.}

\item{file_path}{Included only for API consistency; this must be NULL.}

\item{bbox}{Optionally, an sf::st_bbox() object, or an object that can be converted to such.}

\item{api}{Included only for API consistency; this must be TRUE.}
}
\value{
A library(sf) enabled dataframe comprising perimeters of current wildfires.
}
\description{
Acquire wildfire perimeters
}
\examples{
\dontrun{
get_current_fire_perimeters()
}
}
