% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_dataset_columns}
\alias{get_dataset_columns}
\title{Get the raw column names for a specified dataset}
\usage{
get_dataset_columns(dataset)
}
\arguments{
\item{dataset}{The name of the dataset. One of c('nfip_policies', 'ihp_registrations').}
}
\value{
A vector of raw column names to be selected from the specified dataset
}
\description{
Get the raw column names for a specified dataset
}
