% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_disaster_dollar_database.R
\name{get_disaster_dollar_database}
\alias{get_disaster_dollar_database}
\title{Get Disaster Dollar Database Data}
\usage{
get_disaster_dollar_database(
  file_path = file.path(get_box_path(), "hazards", "carnegie-endowment",
    "disaster_dollar_database_2025_11_19.csv")
)
}
\arguments{
\item{file_path}{The path (on Box) to the file containing the raw data.}
}
\value{
A dataframe comprising disaster-level observations with financial assistance metrics from FEMA's Individual and Households Program (IHP), Public Assistance (PA), HUD's Community Development Block Grant Disaster Recovery (CDBG-DR), and SBA disaster loans.
}
\description{
Get Disaster Dollar Database Data
}
\details{
These data are sourced from: https://carnegieendowment.org/features/disaster-dollar-database. The data returned from this function are unchanged, though some columns have been renamed slightly for clarity and consistency.
}
\examples{
\dontrun{
get_disaster_dollar_database()
}
}
