% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_emergency_managerment_performance.R
\name{get_emergency_management_performance}
\alias{get_emergency_management_performance}
\title{Get EMPG data}
\usage{
get_emergency_management_performance(
  file_path = file.path(get_box_path(), "hazards", "FEMA",
    "emergency-management-performance",
    "emergency_management_performance_grants_2025_06_29.csv"),
  api = TRUE
)
}
\arguments{
\item{file_path}{Path to the downloaded dataset on Box.}

\item{api}{Logical indicating whether to use the OpenFEMA API to retrieve the data. Default is TRUE.}
}
\value{
A data frame containing emergency management performance grant (EMPG) data.
}
\description{
Get EMPG data
}
