% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fema_disaster_declarations.R
\name{get_fema_disaster_declarations}
\alias{get_fema_disaster_declarations}
\title{Get major disaster declarations by county}
\usage{
get_fema_disaster_declarations(
  file_path = file.path(get_box_path(), "hazards", "fema", "disaster-declarations",
    "raw", "fema_disaster_declarations_county_2024_10_25.csv"),
  api = TRUE
)
}
\arguments{
\item{file_path}{The path (on Box) to the file containing the raw data.}

\item{api}{If TRUE (default), access data from the API. Else, read locally from \code{file_path}.}
}
\value{
A dataframe comprising Major Disaster Declarations by month by year by county. Tribal declarations are stored as an attribute of the primary dataframe called \code{tribal_declarations}.
}
\description{
Get major disaster declarations by county
}
\examples{
\dontrun{
get_fema_disaster_declarations(api = TRUE)
}
}
