% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_geography_metadata}
\alias{get_geography_metadata}
\title{Get geography metadata about states or counties}
\usage{
get_geography_metadata(geography_type = c("state", "county"), year = 2023)
}
\arguments{
\item{geography_type}{One of c("state", "county").}

\item{year}{The year for which to obtain state/county metadata. Cannot be greater than the most recent year supported by \code{library(tidycensus)} for the 5-year ACS.}
}
\value{
A data frame containing metadata about the specified geography type and area.
}
\description{
Get geography metadata about states or counties
}
