% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_government_finances.R
\name{get_government_finances}
\alias{get_government_finances}
\title{Get government unit-level expenses from the Census of Governments}
\usage{
get_government_finances(year = 2022)
}
\arguments{
\item{year}{A four-digit year. The default is 2022.}
}
\value{
A dataframe containing government unit-level expenses for the specified year.
}
\description{
Get government unit-level expenses from the Census of Governments
}
