% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ihp_registrations.R
\name{get_ihp_registrations}
\alias{get_ihp_registrations}
\title{Get Individuals and Households Program (IHP) registrations}
\usage{
get_ihp_registrations(
  geography = NULL,
  state_fips,
  file_name = "IndividualsAndHouseholdsProgramValidRegistrations_2024_10_13.parquet",
  api = TRUE,
  outpath = NULL
)
}
\arguments{
\item{geography}{Included only for API consistency; this must be NULL.}

\item{state_fips}{A character vector of two-digit state codes.}

\item{file_name}{The name (not the full path) of the Box file containing the raw data.}

\item{api}{If TRUE (default), query the API.}

\item{outpath}{The path to save the parquet-formatted datafile. Applicable only when \code{api = FALSE}.}
}
\value{
A dataframe comprising county-level data on current NFIP policies
}
\description{
Get Individuals and Households Program (IHP) registrations
}
\examples{
\dontrun{
get_ihp_registrations(
   county_geoids = c("01001", "01003"),
   api = TRUE)
}
}
