% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sba_loans.R
\name{get_sba_loans}
\alias{get_sba_loans}
\title{Access SBA data on disaster loans}
\usage{
get_sba_loans()
}
\value{
A dataframe comprising city- and zip-level data on SBA loanmaking
}
\description{
Access SBA data on disaster loans
}
\examples{
\dontrun{
get_sba_loans()
}
}
