% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sheldus.R
\name{get_sheldus}
\alias{get_sheldus}
\title{Access temporal county-level SHELDUS hazard counts.}
\usage{
get_sheldus(
  file_path = file.path(get_box_path(), "hazards", "sheldus",
    "SHELDUS_22.0_01012916_01012023_AllCounties_CountyAggregate_YearMonth_2022USD",
    "UID14148f_AGG_A.csv")
)
}
\arguments{
\item{file_path}{The path to the raw SHELDUS data.}
}
\value{
A dataframe comprising month x year x county observations of hazard events.
}
\description{
Access temporal county-level SHELDUS hazard counts.
}
\examples{
\dontrun{
get_sheldus()
}
}
