% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sheldus.R
\name{get_sheldus}
\alias{get_sheldus}
\title{Access temporal county-level SHELDUS hazard damage data.}
\usage{
get_sheldus(
  file_path = file.path(get_box_path(), "hazards", "sheldus",
    "SHELDUS_23.0_12312023_AllCounties_CountyAggregate_YearMonthHazard_2023USD",
    "direct_loss_aggregated_output_24075.csv")
)
}
\arguments{
\item{file_path}{The path to the raw SHELDUS data.}
}
\value{
A dataframe comprising hazard x month x year x county observations of hazard events.
}
\description{
Access temporal county-level SHELDUS hazard damage data.
}
\examples{
\dontrun{
get_sheldus()
}
}
