% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_spatial_extent_census}
\alias{get_spatial_extent_census}
\title{Get the Census geographies that overlap with the input spatial dataset}
\usage{
get_spatial_extent_census(data, return_geometry = FALSE)
}
\arguments{
\item{data}{An sf-formatted dataframe}

\item{return_geometry}{Logical. Include the geometries of returned geographies?}
}
\value{
A dataframe (optionally, an sf-dataframe) comprising Census geographies
}
\description{
Get the Census geographies that overlap with the input spatial dataset
}
