% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_structures.R
\name{get_structures}
\alias{get_structures}
\title{Estimate counts of hazard-impacted structures by structure type}
\usage{
get_structures(boundaries, geography = "county", keep_structures = FALSE)
}
\arguments{
\item{boundaries}{A POLYGON or MULTIPOLYGON object, or an sf::st_bbox()-style bbox.}

\item{geography}{The desired geography of the results. One of "tract" or "county".}

\item{keep_structures}{Logical. If TRUE, the raw structure data will be returned alongside the summarized data.}
}
\value{
A dataframe comprising estimated counts of each structure type, at the specified \code{geography}, for all such geographic units intersecting the \code{boundaries} object. If keep_structure = TRUE, returns a list with two elements: the summarized data and the raw structure data.
}
\description{
Estimate counts of hazard-impacted structures by structure type
}
\examples{
\dontrun{
get_structures(
  geography = "tract",
  boundaries = tigris::states(cb = TRUE) \%>\% dplyr::filter(stringr::str_detect(NAME, "District")))
}
}
