% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_system_username}
\alias{get_system_username}
\title{Get the user's username}
\usage{
get_system_username()
}
\value{
The username of the user running the script
}
\description{
Get the user's username
}
\examples{
\dontrun{
get_system_username()
}
}
