% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{inflation_adjust}
\alias{inflation_adjust}
\title{Inflation adjust dollar values using annual PCE Index}
\usage{
inflation_adjust(
  df,
  year_variable,
  dollar_variables,
  names_suffix = NULL,
  base_year = 2024
)
}
\arguments{
\item{df}{The dataframe with values to inflation-adjust}

\item{year_variable}{The name of the column denoting the year that existing dollar-denominated figures are based in.}

\item{dollar_variables}{The variables to inflation-adjust.}

\item{names_suffix}{A suffix to add to the names of the inflation-adjusted variables. If NULL, defaults to "_<base_year>". If "", columns are renamed in place.}

\item{base_year}{The year to use as the base for inflation adjustment. If NULL, defaults to the most recent year in the PCE index data.}
}
\value{
A dataframe with inflation-adjusted values
}
\description{
The Personal Consumption Expenditures Price Index (PCE Index) is from the
Federal Reserve Bank of St. Louis's FRED tool.
}
\examples{
\dontrun{
df = tibble::tribble(
  ~ year, ~ amount,
  1990, 1,
  1991, 1,
  1992, 1)

df |>
  inflation_adjust(
    year_variable = year,
    dollar_variables = amount,
    names_suffix = "inflation_adjusted")
}
}
