% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics_analysis.R
\name{qualtrics_define_missing}
\alias{qualtrics_define_missing}
\title{Fill in missing and non-missing values across interrelated survey questions}
\usage{
qualtrics_define_missing(
  df,
  question_code_include,
  question_code_omit = NULL,
  default_values = list("No", 0, as.Date(0)),
  predicate_question = NULL,
  predicate_question_negative_value = NULL
)
}
\arguments{
\item{df}{A dataframe of survey responses}

\item{question_code_include}{A regex that matches the columns to include in the missing non-missing value imputation}

\item{question_code_omit}{A regex that matches the columns to omit from the missing non-missing value imputation}

\item{default_values}{A list of length three, specifying the default, non-missing values to be used for character, numeric, and Date columns, respectively}

\item{predicate_question}{Optional. The name of a single column that controls whether columns selected with \code{question_code_include}}

\item{predicate_question_negative_value}{If \code{predicate_question} is specified, provide the value that indicates a negative response to the predicate question. For responses where the predicate question has this value, this value will be imputed to the specified columns}
}
\value{
The inputted \code{df} object with missing/non-missing values applied to specified columns
}
\description{
Fill in missing and non-missing values across interrelated survey questions
}
