% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics_analysis.R
\name{qualtrics_plot_question}
\alias{qualtrics_plot_question}
\title{Plot responses to Qualtrics survey questions}
\usage{
qualtrics_plot_question(
  df,
  metadata,
  question_code_include,
  question_code_omit = "zzzzz",
  question_type,
  title = "",
  subtitle = NULL,
  subtitle_replace = c(`\\\\[Field.*\\\\]` = "your community",
    `Your best estimate is fine\\\\.` = ""),
  text_remove = "please describe|please specify",
  text_replace = c(a = "a"),
  omit_other = TRUE
)
}
\arguments{
\item{df}{A dataframe of survey responses}

\item{metadata}{A dataframe of Qualtrics metadata}

\item{question_code_include}{A regex that matches the question codes to include in the plot}

\item{question_code_omit}{A regex that matches the question codes to omit from the plot}

\item{question_type}{one of c("checkbox_single", "checkbox_multi", "checkbox_factor")}

\item{title}{Plot title}

\item{subtitle}{Plot subtitle}

\item{subtitle_replace}{A named character vector of regex patterns to replace in the subtitle}

\item{text_remove}{A regex pattern to select response options to exclude from the plot}

\item{text_replace}{A named character vector of regex patterns to replace in the response text}

\item{omit_other}{Logical; whether to omit the "Other" response option. Default is TRUE.}
}
\value{
A ggplot object
}
\description{
Plot responses to Qualtrics survey questions
}
