% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{read_xlsx_from_url}
\alias{read_xlsx_from_url}
\title{Download a .xlsx file(s) from a URL(s)}
\usage{
read_xlsx_from_url(urls, directory, file_names = NULL, silent = TRUE)
}
\arguments{
\item{urls}{A character vector of URLs of length one or greater.}

\item{directory}{The path to a single directory--not to a file--where the .xlsx file(s) will be saved.}

\item{file_names}{Optionally, a character vector of the same length as \code{urls} containing only the file names (not the full paths) with which the downloaded files should be named. If NULL (default), file names are extracted from \code{urls}.}

\item{silent}{If TRUE (default), files are saved silently. If FALSE, downloaded files are read and returned as a list.}
}
\value{
Either nothing (silent == TRUE) or a list of dataframes from the specified URLs.
}
\description{
Download a .xlsx file(s) from a URL(s)
}
